refactor: 경로 조회와 요금 조회 기능 분리
