+++ src/main/java/wooteco/subway/path/dto/PathResponseAssembler.java
@@ -1,5 +1,6 @@
 package wooteco.subway.path.dto;
 
+import wooteco.subway.path.domain.SubwayFare;
 import wooteco.subway.path.domain.SubwayPath;
 import wooteco.subway.station.dto.StationResponse;
 
@@ -8,12 +9,13 @@ import java.util.stream.Collectors;
 
 public class PathResponseAssembler {
     public static PathResponse assemble(SubwayPath subwayPath, int age) {
+        SubwayFare subwayFare = new SubwayFare(subwayPath.getSectionEdges());
         List<StationResponse> stationResponses = subwayPath.getStations().stream()
                 .map(it -> StationResponse.of(it))
                 .collect(Collectors.toList());
 
         int distance = subwayPath.calculateDistance();
-        int fare = subwayPath.calculateFare(distance, age);
+        int fare = subwayFare.calculateFare(distance, age);
 
         return new PathResponse(stationResponses, distance, fare);
     }
